+++ valib/filters/decoder_graph.h
@@ -32,7 +32,7 @@ public:
   Speakers get_user() const;
 
   // Summary information
-  int get_info(char *_buf, size_t _len) const;
+  size_t get_info(char *_buf, size_t _len) const;
 
   /////////////////////////////////////////////////////////////////////////////
   // Filter overrides
