+++ valib/filters/proc.h
@@ -118,7 +118,7 @@ public:
   /////////////////////////////////////////////////////////
   // AudioProcessior interface
 
-  int get_info(char *buf, size_t len) const;
+  size_t get_info(char *buf, size_t len) const;
 
   size_t get_buffer() const;
   void   set_buffer(size_t nsamples);
